+++ src/strict-rules.ts
@@ -74,6 +74,7 @@ export const strictRules: DummyRuleMap = {
   ],
   "typescript/use-unknown-in-catch-callback-variable": "error",
   "unicorn/no-accessor-recursion": "error",
+  "unicorn/no-array-fill-with-reference-type": "error",
   "oxc/double-comparisons": "off",
   "oxc/erasing-op": "off",
   "oxc/number-arg-out-of-range": "off",